+++ src/fastify.d.ts
@@ -4,4 +4,8 @@ declare module 'fastify' {
   export interface FastifyInstance {
     pg: PostgresDb;
   }
+
+  export interface FastifyRequest {
+    slowTimer: ReturnType<typeof setTimeout> | undefined;
+  }
 }